+++ test/file-read-perf.js
@@ -57,7 +57,7 @@ function makeTest (timerLabel, numFiles) {
     const deploy = require('../commands/deploy')
     const upload = (dir) => {
       console.time(timerLabel)
-      deploy([dir], {})
+      deploy([dir], { cwd: '.' })
     }
 
     let tempPath